refactor(web): share the event card between the events page and the calendar
Move the event card out of the events route into the dashboard feature
so the calendar pane can render the same card. No visual change.

The one API change: whether the event is past is now a prop rather than
read from the clock inside the card. The card is memoised, so a clock
read inside it would freeze at first render; the parent re-renders on
whatever tick it already has (the events list on a data refresh, the
week grid's today column on its minute tick) and passes the answer in.
